Some changes to gstomx.conf handling
A few tweaks to the sed rules to make more friendly for using macros (for
example, "foo" "bar" becomes "foobar", which is normally done by the C
compiler and not the pre-processor).  Also, the result of the preprocessor,
before generating gstomx_conf.c, is now also a generated file, to ease
debugging and to use as a starting point for a customized config file.

Signed-off-by: Rob Clark <rob@ti.com>