base_filter: remove memory copying when the buffers are output from gst-openmax
Instead of using memcpy to copy the output buffer contents, send the
pre-allocated buffers to the sink plugin.

The pre-allocated buffers are registered with UIOMUX and their physical
address can be obtained by OMX_AllocateBuffer() for output port.
This allows libshvio to convert the frame data without memory copying.

The current implementation doesn't keep track of any reference counts on
the buffers handed out to the sink plugin.  As such, if the sink operates
asychronously with respect to the gst-openmax plugin, additional care will
need to be taken to ensure that the sink is finished with the buffer before
the gst-openmax plugin tries to use it again.