base_filter: set proper data size to a GstBuffer instance when the post processing is disabled
In the post processing disabled, nFilledLen doesn't contain actual
data size but the image size assumed to be transformed with the post
processing. To figure out the actual size, we should calculate it
from nStride and nSliceHeight that are obtained from
OMX_VIDEO_PORTDEFINITIONTYPE. chroma_byte_offset has a value that is
nStride * nSliceHeight. The total data size equals chroma_byte_offset * 3 / 2.