util: change the beginning of locked section to definitely wakeup blocking a thread when an error occurs
When OMX_EventError is caught after OMX_SendCommand() is called and before
g_cond_timed_wait() is invoked, the thread blocks until timeout because the
wakeup signal from the event handler is sent before the thread blocks.
This change moves the mutex lock to before OMX_SendCommand() in order to
avoid this race condition.