util: don't send an omx_state_condition signal when catching an error unrelated to changing state
This signal is used in order to wakeup g_cond_timed_wait() that waits
for state change completion. When changing state fails, the signal
should be sent to force the blocking thread to wakeup. However,
the signal is sent when any error occurs, regardless of whether it
was related to changing state or not.

This patch adds a test to determine if an error is caused by changing
state so that the signal is only send when appropriate.
This test only looks at the error code, and so may cause a false
positive if an error that could be caused by a state change occurs
for some other reason.

The errors that correspond to state changes were determined by
using the REL OMXIL user manual ver.1.1 as a reference.