util/base_filter: send an error event to application immediately after catching an error from the component
An error event might not be sent to application after catching an error
from the OMXIL component because the error event isn't sent when
EventHandler() is invoked but is sent after pad_chain() detects it.
This patch notifies the application of the error from the component
in the event handler in order to ensure that the the error is sent.