mpeg4dec: specify the video stream type as packed bitstream
REL OMXIL can be set the video stream type, which is one of the
extended parameters. This setting is necessary to decode mpeg4 streams
with packed bitstreams. Non-packed bitstreams can be decoded regardless
of the value of this setting, so it is safe to simply turn it on for
all mpeg4 streams.
This patch overrides base omx_setup() method of mpeg4dec
in order to set this param.