mpeg4dec: concatenate codec_data with the first frame
When a mpeg4 stream is demuxed, codec_data may have some
decoder configuration information (DCI). The DCI contains
some data that is necessary for decoding the stream that follows it.
According to the REL OMXIL spec, the DCI must be concatenated with the
first frame and pushed with EmptyThisBuffer() when the packed bitstream
is specified.
The DCI information is appended even if the same data already exists
within the first frame, since the OMXIL will silently discard the
duplicate DCI data.